pyapi/requirements.txt
fastapi
uvicorn
requests
tqdm
kaggle
python-dateutil
certifi
urllib3
idna
chardet
python-slugify
pyyaml
tabulate
bleach
webencodings
html5lib
unidecode
google
google-auth
google-auth-oauthlib
google-api-python-client
protobuf
six
pillow
python-multipart
httpx
pymysql
gunicorn
bloom-filter2
mysql-connector-python
python-dotenv
